Split RolesManagerPanel into a tab shell and extracted OrbatRolesTab
